updating annotations: if the current user is not owner then the permission part is ignored
git-svn-id: https://svn.clarin.eu/DASISH/t5.6/backend/annotator-backend/trunk/annotator-backend@5774 72633d89-856b-0410-a642-59178932f03a

+++ src/main/java/eu/dasish/annotation/backend/dao/impl/DBDispatcherImlp.java
@@ -696,14 +696,20 @@ public class DBDispatcherImlp implements DBDispatcher {
 // TODO: unit test
 
     @Override
-    public int updateAnnotation(Annotation annotation) throws NotInDataBaseException {
+    public int updateAnnotation(Annotation annotation, String remoteUser) throws NotInDataBaseException {
         Number annotationID = annotationDao.getInternalID(UUID.fromString(annotation.getId()));
-        int updatedAnnotations = annotationDao.updateAnnotation(annotation, annotationID, principalDao.getInternalIDFromHref(annotation.getOwnerHref()));
+        Number ownerID = principalDao.getInternalIDFromHref(annotation.getOwnerHref());
+        int updatedAnnotations = annotationDao.updateAnnotation(annotation, annotationID, ownerID);
         int deletedTargets = annotationDao.deleteAllAnnotationTarget(annotationID);
-        int deletedPrinsipalsAccesss = annotationDao.deleteAnnotationPermissions(annotationID);
         int addedTargets = this.addTargets(annotation, annotationID);
-        int addedPrincipalsAccesss = this.addPermissions(annotation.getPermissions().getPermission(), annotationID);
-        int updatedPublicAttribute = annotationDao.updatePublicAttribute(annotationID, annotation.getPermissions().getPublic());
+        
+        Number remoteUserID = principalDao.getPrincipalInternalIDFromRemoteID(remoteUser);
+        
+        if (ownerID.equals(remoteUserID)) {
+            int deletedPrinsipalsAccesss = annotationDao.deleteAnnotationPermissions(annotationID);
+            int addedPrincipalsAccesss = this.addPermissions(annotation.getPermissions().getPermission(), annotationID);
+            int updatedPublicAttribute = annotationDao.updatePublicAttribute(annotationID, annotation.getPermissions().getPublic());
+        };
         return updatedAnnotations;
     }
